refactor: constants 컨벤션 적용
