+++ frontend/src/constants/index.ts
@@ -2,3 +2,6 @@ export * from '@/constants/keys';
 export * from '@/constants/oauths';
 export * from '@/constants/cropper';
 export * from '@/constants/time';
+export * from '@/constants/location';
+export * from '@/constants/style';
+export * from '@/constants/walk';
